Don't mix ESM and CommonJS in shim/dynamic.js
It seems that at some point when this was written someone mixed ESM and
CommonJS syntax.

There is not good reason to do this, and it will soon not be supported -
and arguably has never been, it just happened to work due to how ESM was
supported in k6.

See https://github.com/grafana/k6/pull/3807 and linked issues/PRs